backend/requirements.txt
fastapi==0.88.0
uvicorn==0.20.0
python-dotenv==1.0.1
requests==2.31.0
pydantic==1.10.13
python-multipart==0.0.9
websockets==12.0
elevenlabs==1.54.0
pytest==8.0.2
httpx==0.27.0
aiosqlite==0.19.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0.post1 